requirements.txt
cassandra-driver == 3.19.0
requests